load very requirements from the requirements class (2.1.8)
